+++ apps/vs-agent/src/controllers/public/index.ts
@@ -1,3 +1,3 @@
 export * from './didcomm'
 export * from './invitation'
-export * from './self-vtr'
+export * from './self-tr'
