+++ apps/vs-agent/src/controllers/public/self-tr/index.ts
@@ -1,0 +1,2 @@
+export * from './SelfTrController'
+export * from './SelfTrService'
