--- apps/vs-agent/src/controllers/public/self-vtr/index.ts
@@ -1,2 +1,0 @@
-export * from './SelfVtrController'
-export * from './SelfVtrService'
