+++ apps/vs-agent/src/utils/data.ts
@@ -1,7 +1,9 @@
-export function getEcsSchemas(publicApiBaseUrl: string) {
+import { AnySchemaObject } from 'ajv'
+
+export function getEcsSchemas(publicApiBaseUrl: string): AnySchemaObject {
   return {
     'ecs-org': {
-      $id: `${publicApiBaseUrl}/self-tr/v1/cs/js/ecs-org`,
+      $id: `${publicApiBaseUrl}/self-tr/cs/v1/js/ecs-org`,
       $schema: 'https://json-schema.org/draft/2020-12/schema',
       title: 'OrganizationCredential',
       description: 'OrganizationCredential using JsonSchema',
@@ -54,7 +56,7 @@ export function getEcsSchemas(publicApiBaseUrl: string) {
       },
     },
     'ecs-person': {
-      $id: `${publicApiBaseUrl}/self-tr/v1/cs/js/ecs-person`,
+      $id: `${publicApiBaseUrl}/self-tr/cs/v1/js/ecs-person`,
       $schema: 'https://json-schema.org/draft/2020-12/schema',
       title: 'PersonCredential',
       description: 'PersonCredential using JsonSchema',
@@ -97,7 +99,7 @@ export function getEcsSchemas(publicApiBaseUrl: string) {
       },
     },
     'ecs-service': {
-      $id: `${publicApiBaseUrl}/self-tr/v1/cs/js/ecs-service`,
+      $id: `${publicApiBaseUrl}/self-tr/cs/v1/js/ecs-service`,
       $schema: 'https://json-schema.org/draft/2020-12/schema',
       title: 'ServiceCredential',
       description: 'ServiceCredential using JsonSchema',
@@ -166,7 +168,7 @@ export function getEcsSchemas(publicApiBaseUrl: string) {
       },
     },
     'ecs-user-agent': {
-      $id: `${publicApiBaseUrl}/self-tr/v1/cs/js/ecs-user-agent`,
+      $id: `${publicApiBaseUrl}/self-tr/cs/v1/js/ecs-user-agent`,
       $schema: 'https://json-schema.org/draft/2020-12/schema',
       title: 'UserAgentCredential',
       description: 'UserAgentCredential using JsonSchema',
